move test settings, helper functions, fixtures to separate file
